flask_server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
google-generativeai
python-dotenv
langchain
langchain-google-genai
